requirements.txt
torch
transformers
fpdf
streamlit
PyMuPDF
pytesseract
Pillow